--- a/helpers.go
+++ b/helpers.go
@@ -10,6 +10,9 @@ type ForEachFn func(key, val []byte) (end bool)
 // GrowFn is used when calling grow internally
 type GrowFn func(sz int64) (bs []byte)
 
+// CloseFn is used when close is called
+type CloseFn func() error
+
 func getSize(cnt, keySize, valSize int64) (sz int64) {
 	sz = keySize + valSize + blockSize
 	sz *= cnt