--- a/helpers.go
+++ b/helpers.go
@@ -4,6 +4,13 @@ type color uint8
 
 type childType uint8
 
+type trunk struct {
+	root int64
+	cnt  int64
+	tail int64
+	cap  int64
+}
+
 // ForEachFn is used when calling ForEach from a Tree
 type ForEachFn func(key, val []byte) (end bool)