For the most part, the performance is identical. This commit greatly cleans up the backend system. If there are any suggestions, I am all ears!

```bash
name                    old time/op    new time/op    delta
RBTGet-16                 3.39ms ± 1%    3.41ms ± 3%    ~     (p=0.548 n=5+5)
RBTSortedGetPut-16        6.94ms ± 2%    7.01ms ± 2%    ~     (p=0.421 n=5+5)
RBTSortedPut-16           3.73ms ± 2%    3.70ms ± 2%    ~     (p=0.690 n=5+5)
RBTReversePut-16          3.69ms ± 2%    3.74ms ± 2%    ~     (p=0.421 n=5+5)
RBTRandomPut-16           4.09ms ± 2%    4.13ms ± 0%  +0.81%  (p=0.032 n=5+4)
RBTForEach-16             93.9µs ± 1%    93.9µs ± 1%    ~     (p=1.000 n=5+5)
RBTMMapGet-16             3.42ms ± 2%    3.43ms ± 2%    ~     (p=0.421 n=5+5)
RBTMMapSortedGetPut-16    6.93ms ± 2%    7.03ms ± 3%    ~     (p=0.222 n=5+5)
RBTMMapSortedPut-16       3.69ms ± 3%    3.76ms ± 2%    ~     (p=0.222 n=5+5)
RBTMMapReversePut-16      3.68ms ± 2%    3.76ms ± 2%    ~     (p=0.056 n=5+5)
RBTMMapRandomPut-16       4.35ms ± 3%    4.39ms ± 2%    ~     (p=0.421 n=5+5)
RBTMMapForEach-16         93.1µs ± 2%    93.8µs ± 1%    ~     (p=0.421 n=5+5)

name                    old alloc/op   new alloc/op   delta
RBTGet-16                  0.00B          0.00B         ~     (all equal)
RBTSortedGetPut-16         0.00B          0.00B         ~     (all equal)
RBTSortedPut-16            0.00B          0.00B         ~     (all equal)
RBTReversePut-16           0.00B          0.00B         ~     (all equal)
RBTRandomPut-16            0.00B          0.00B         ~     (all equal)
RBTForEach-16              0.00B          0.00B         ~     (all equal)
RBTMMapGet-16              0.00B          0.00B         ~     (all equal)
RBTMMapSortedGetPut-16     0.00B          0.00B         ~     (all equal)
RBTMMapSortedPut-16        0.00B          0.00B         ~     (all equal)
RBTMMapReversePut-16       0.00B          0.00B         ~     (all equal)
RBTMMapRandomPut-16        0.00B          0.00B         ~     (all equal)
RBTMMapForEach-16          0.00B          0.00B         ~     (all equal)

name                    old allocs/op  new allocs/op  delta
RBTGet-16                   0.00           0.00         ~     (all equal)
RBTSortedGetPut-16          0.00           0.00         ~     (all equal)
RBTSortedPut-16             0.00           0.00         ~     (all equal)
RBTReversePut-16            0.00           0.00         ~     (all equal)
RBTRandomPut-16             0.00           0.00         ~     (all equal)
RBTForEach-16               0.00           0.00         ~     (all equal)
RBTMMapGet-16               0.00           0.00         ~     (all equal)
RBTMMapSortedGetPut-16      0.00           0.00         ~     (all equal)
RBTMMapSortedPut-16         0.00           0.00         ~     (all equal)
RBTMMapReversePut-16        0.00           0.00         ~     (all equal)
RBTMMapRandomPut-16         0.00           0.00         ~     (all equal)
RBTMMapForEach-16           0.00           0.00         ~     (all equal)
```